% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkOrgsRestoreML.R
\name{RestoreML}
\alias{RestoreML}
\title{RestoreML}
\usage{
RestoreML(...)
}
\description{
A function, primarily for internal used, used to initialize the machine learning models used in the record linkage algorithms of Libgober and Jerzak.
}
